requirements.txt
# ASR stack
whisperx==3.7.4
pyannote.audio==3.4.0
omegaconf>=2.3.0,<3.0.0

# Telegram bot
python-telegram-bot==21.6
httpx~=0.27

# --- GPT / Protocol ---
openai>=1.30.0

# --- Word export (protocol.docx) ---
python-docx>=1.1.0